Payment confirmation redirect

+++ src/databases/migrations/2020_02_27_024014_create_interswitch_payments_table.php
@@ -21,7 +21,7 @@ class CreateInterswitchPaymentsTable extends Migration
             $table->string('environment');
             $table->string('gateway');
             $table->string('reference');
-            $table->float('amount'); // always store in kobo
+            $table->bigInteger('amount'); // always store in kobo
             $table->string('response_code')->nullable();
             $table->string('response_description')->nullable();
             $table->timestamps();
